+++ login.html
@@ -47,10 +47,13 @@
                 placeholder="비밀번호를 입력해주세요."
                 required
               />
-              <img class="eye-image" src="./images/login/password_close_eye_btn.svg" />
+              <img
+                class="eye-image"
+                src="./images/login/password_close_eye_btn.svg"
+              />
             </div>
           </div>
-          <span class="error-text err-passwordConfirm"></span>
+          <span class="error-text err-password"></span>
           <button class="go-login" type="submit" disabled>로그인</button>
         </form>
 
@@ -74,6 +77,6 @@
         </div>
       </div>
     </div>
-    <script src="./javaScript/login.js"></script>
+    <script src="./javaScript/login-signup/login.js"></script>
   </body>
 </html>
