+++ signup.html
@@ -59,7 +59,10 @@
                 placeholder="비밀번호를 입력해주세요."
                 required
               />
-              <img class="eye-image" src="./images/login/password_close_eye_btn.svg" />
+              <img
+                class="eye-image"
+                src="./images/login/password_close_eye_btn.svg"
+              />
             </div>
             <span class="error-text err-password"></span>
           </div>
@@ -74,7 +77,10 @@
                 placeholder="비밀번호를 입력해주세요."
                 required
               />
-              <img class="eye-image" src="./images/login/password_open_eye_btn.svg" />
+              <img
+                class="eye-image"
+                src="./images/login/password_open_eye_btn.svg"
+              />
             </div>
             <span class="error-text err-password_confirm"></span>
           </div>
@@ -101,6 +107,6 @@
         </div>
       </div>
     </div>
-    <script src="./javaScript/login.js"></script>
+    <script src="./javaScript/login-signup/singup.js"></script>
   </body>
 </html>
